# Back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Lagafy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api/",
    "paths": {
        "/reservation/myreservations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Get my reservations list.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Reservation"
                            }
                        }
                    }
                }
            }
        },
        "/reservation/status/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Make Check-In of user for the selected travel.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Travel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Reservation"
                        }
                    }
                }
            }
        },
        "/reservation/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Delete reservation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id reservation",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/reservation/{travelId}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Create reservation on the given travel.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Travel",
                        "name": "travelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Reservation"
                        }
                    }
                }
            }
        },
        "/travel/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "travels"
                ],
                "summary": "Create new travel",
                "parameters": [
                    {
                        "description": "Create a travel",
                        "name": "travel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Travel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Travel"
                        }
                    }
                }
            }
        },
        "/travel/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "travels"
                ],
                "summary": "Get travel by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "travel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Travel"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "travels"
                ],
                "summary": "Update travel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id travel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a travel",
                        "name": "travel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Travel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Travel"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "travels"
                ],
                "summary": "Delete travel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "travel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/travels/current": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "travels"
                ],
                "summary": "Get currents travels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Travel"
                            }
                        }
                    }
                }
            }
        },
        "/travels/information": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "travels"
                ],
                "summary": "Get information from travels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date. Expample: 2020-02-20",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date. Expample: 2020-03-10",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TravelInformation"
                            }
                        }
                    }
                }
            }
        },
        "/travels/templates": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "travels"
                ],
                "summary": "Get templates for travels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Travel"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Passenger": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Reservation": {
            "type": "object",
            "required": [
                "passenger"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "passenger": {
                    "type": "object",
                    "$ref": "#/definitions/models.Passenger"
                },
                "passengerID": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "travel": {
                    "type": "object",
                    "$ref": "#/definitions/models.Travel"
                },
                "travelID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Travel": {
            "type": "object",
            "required": [
                "arrivalAddress",
                "capacity",
                "departure",
                "departureAddress",
                "destination",
                "origin"
            ],
            "properties": {
                "arrival": {
                    "type": "string"
                },
                "arrivalAddress": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "departure": {
                    "type": "string"
                },
                "departureAddress": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "driver": {
                    "type": "string"
                },
                "hasWifi": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "origin": {
                    "type": "string"
                },
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Reservation"
                    }
                },
                "status": {
                    "type": "string"
                },
                "template": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TravelInformation": {
            "type": "object",
            "required": [
                "capacity",
                "departure",
                "destination",
                "origin"
            ],
            "properties": {
                "arrival": {
                    "type": "string"
                },
                "cancelled": {
                    "type": "integer"
                },
                "capacity": {
                    "type": "integer"
                },
                "confirmed": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "departure": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "onboard": {
                    "type": "integer"
                },
                "origin": {
                    "type": "string"
                },
                "pending": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}